+++ test/common/ssl/test_data/no_san_cert_info.h
@@ -1,7 +1,7 @@
 // NOLINT(namespace-envoy)
 constexpr char TEST_NO_SAN_CERT_HASH[] =
-    "13a18d87a20d47d48fc780f72f13c7229a7cd11a4580480c3ee7103e4441a2ed";
-constexpr char TEST_NO_SAN_CERT_SPKI[] = "pzOag3sb9kavj7MJDYlS+f7vlA16i0EVuLm5rjpf8FY=";
-constexpr char TEST_NO_SAN_CERT_SERIAL[] = "97b12a171d4a657f";
-constexpr char TEST_NO_SAN_CERT_NOT_BEFORE[] = "Dec 16 07:33:53 2018 GMT";
-constexpr char TEST_NO_SAN_CERT_NOT_AFTER[] = "Dec 15 07:33:53 2020 GMT";
+    "3a27b98c0eaf0d9162740402bab20a0551422a6fe36b7f982a4096e2fd2a2a92";
+constexpr char TEST_NO_SAN_CERT_SPKI[] = "J5LodXBW3zUhY78+P2fCYDvsDj/L8C7sDNiiImj27HA=";
+constexpr char TEST_NO_SAN_CERT_SERIAL[] = "8a80523e06a57565";
+constexpr char TEST_NO_SAN_CERT_NOT_BEFORE[] = "Dec 17 20:17:25 2018 GMT";
+constexpr char TEST_NO_SAN_CERT_NOT_AFTER[] = "Dec 16 20:17:25 2020 GMT";