+++ test/common/ssl/test_data/san_only_dns_cert_info.h
@@ -1,7 +1,7 @@
 // NOLINT(namespace-envoy)
 constexpr char TEST_SAN_ONLY_DNS_CERT_HASH[] =
-    "e1cbe8764c79afb0366f07122611ca0698e5ac6ff4f134f2ba73a29831b5f227";
-constexpr char TEST_SAN_ONLY_DNS_CERT_SPKI[] = "g+GfEuj1a7VnD0zJfzlk4DoRc52jZNdOBpZjivnK4Ww=";
-constexpr char TEST_SAN_ONLY_DNS_CERT_SERIAL[] = "97b12a171d4a6583";
-constexpr char TEST_SAN_ONLY_DNS_CERT_NOT_BEFORE[] = "Dec 16 07:33:54 2018 GMT";
-constexpr char TEST_SAN_ONLY_DNS_CERT_NOT_AFTER[] = "Dec 15 07:33:54 2020 GMT";
+    "9a9569ba6a17c16d61fcb7ecaec3362252645ab7d59f8784811e11ef2c4df814";
+constexpr char TEST_SAN_ONLY_DNS_CERT_SPKI[] = "gh04a2Iuqy70cLueDJBxsO14xlSI6rppzxDXgvjQWds=";
+constexpr char TEST_SAN_ONLY_DNS_CERT_SERIAL[] = "8a80523e06a57569";
+constexpr char TEST_SAN_ONLY_DNS_CERT_NOT_BEFORE[] = "Dec 17 20:17:26 2018 GMT";
+constexpr char TEST_SAN_ONLY_DNS_CERT_NOT_AFTER[] = "Dec 16 20:17:26 2020 GMT";