+++ test/common/ssl/test_data/san_uri_cert_info.h
@@ -1,7 +1,7 @@
 // NOLINT(namespace-envoy)
 constexpr char TEST_SAN_URI_CERT_HASH[] =
-    "e1a998aff8f9ac26dad3fad74a302145201645ee4f87142afb31ce8dd3e0141b";
-constexpr char TEST_SAN_URI_CERT_SPKI[] = "iP/7G36v8SU/4KkX1UvU3NUNFsLxi1dutdyRhPhdwOc=";
-constexpr char TEST_SAN_URI_CERT_SERIAL[] = "97b12a171d4a6584";
-constexpr char TEST_SAN_URI_CERT_NOT_BEFORE[] = "Dec 16 07:33:54 2018 GMT";
-constexpr char TEST_SAN_URI_CERT_NOT_AFTER[] = "Dec 15 07:33:54 2020 GMT";
+    "f214c479742b87363784c953b96d5c496830f1062b640794866b48103e5522ed";
+constexpr char TEST_SAN_URI_CERT_SPKI[] = "X6XHMwcqQzQnup+1TRyqpJfzuMfK6qO5FDLl90o8ahM=";
+constexpr char TEST_SAN_URI_CERT_SERIAL[] = "8a80523e06a5756a";
+constexpr char TEST_SAN_URI_CERT_NOT_BEFORE[] = "Dec 17 20:17:26 2018 GMT";
+constexpr char TEST_SAN_URI_CERT_NOT_AFTER[] = "Dec 16 20:17:26 2020 GMT";